refactor: Separate models into individual files
